requirements.txt
requests
things.py
PyYAML